adding some validation

+++ src/main/java/com/system/library/service/impl/PatronServiceImpl.java
@@ -1,6 +1,9 @@
 package com.system.library.service.impl;
 
+import com.system.library.model.Book;
+import com.system.library.model.BorrowingRecord;
 import com.system.library.model.Patron;
+import com.system.library.repository.BorrowingRecordRepository;
 import com.system.library.repository.PatronRepository;
 import com.system.library.service.PatronService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,6 +16,8 @@ public class PatronServiceImpl implements PatronService {
 
     @Autowired
     private PatronRepository patronRepository;
+    @Autowired
+    private BorrowingRecordRepository borrowingRecordRepository;
 
     @Override
     public List<Patron> getAllPatrons() {
@@ -41,6 +46,18 @@ public class PatronServiceImpl implements PatronService {
 
     @Override
     public void deletePatron(Long id) {
-        patronRepository.deleteById(id);
+        Optional<Patron> patron = patronRepository.findById(id);
+        if (patron.isPresent() && !isBorrowing(id)) {
+            patronRepository.deleteById(id);
+        }else if (isBorrowing(id)) {
+            throw new IllegalStateException("Cannot Delete this patron because he is borrowing a book!");
+        }
+        else {
+            throw new IllegalStateException("No existing patron with id " + id +" to delete");
+        }
+    }
+    private boolean isBorrowing(Long patronId) {
+        Optional<BorrowingRecord> activeBorrowingRecord = borrowingRecordRepository.findFirstByPatronId(patronId);
+        return activeBorrowingRecord.isPresent();
     }
 }
